spacer on bottom, img no longer pushes bottom of page off screen
