+++ src/styles/layout.css
@@ -4,7 +4,6 @@ body {
     font-family: 'Roboto', sans-serif;
     background-color: #171A21;
     color: #7A93AC;
-    
 }
 
 /* KantanReader */
@@ -118,13 +117,14 @@ body {
     display: flex;
     width: 100%;
     height: 100%;
+    max-height: 100vh;
     flex-direction: column;
 }
 
 .readerDisplay {
     display: flex;
     justify-content: space-between;
-    height: calc(100% - 10px);
+    height: calc(100% - 20px);
     width: 100%;
 
 }
@@ -180,4 +180,5 @@ body {
 img {
     max-width: 100%;
     overflow: hidden;
+    max-height: calc(100vh - 60px);
 }
